Queue a task to resolve createImageBitmap
fixes #5329
fixes #10611

We went the "create our own task-queue" way, which seems to be the easiest here.
Now, it's expected that early errors reject synchronously, while everything rejecting or resolving "in parallel" is ran from a task.
This matches Firefox's behavior.
WebKit only matches the synchronous rejections of any source and the async settlement of Blob.
Chrome only matches the synchronous rejections of any source and the async settlement of Blob in a Window context and makes everything async in a Worker.

We also fixed a missing </p> nearby, and ran The Great Rewrapper on the affected paragraph.