Ensure modal dialogs can nest inside of hint popovers
Allow dialogs that are nested inside of a hint popover to be shown as modal. The previous text meant that dialogs may close their parent hint popover causing issues similar to #9998. Make it such that a dialog (or fullscreen element) is a viable candidate to nest inside of a hint popover.

Closes #11008, by making the modifications suggested in that issue.