Editorial: simplify (re)moving steps of form-associated elements
Both the removing steps and moving steps are invoked for all nodes in a removed or moved tree, so the recursion implied by "or the ancestor of" isn't needed.

This better matches implementations, and seems to be unobservable now that mutation events don't exist and no other JavaScript code runs during these steps.