Change definition of list box and drop-down box rendering
This PR allows select elements with the multiple attribute to be
rendered as a drop-down box when the author explicitly sets the size
attribute to 1.

The related issue should be resolved by requiring that mobile browsers
and desktop browsers both follow these rules in order to improve
cross-platform compatibility.

Fixes https://github.com/whatwg/html/issues/8189