Revamp base URL change processing
This removes the dependency on two concepts supposedly defined in DOM, that have not existed for many years. Instead, we have HTML directly call into the new "respond to base URL changes" steps, when mutations happen to a Document that would cause the base URL to change.

Also reorganizes some of the base URL stuff to stick together, instead of being mixed between "Dynamic changes to base URLs" and "Terminology".

Closes https://github.com/whatwg/dom/issues/61.