Clean up Worker and SharedWorker definitions
This especially focuses on the constructor, although it also slightly improves the definitions of SharedWorker's port attribute and Worker's terminate() method.

This is almost entirely editorial, but it does make some fixes (part of #4506) to properly post tasks using "queue a global task" and specifying the task source. (The DOM manipulation task source was chosen, for consistency with the few cases where we did specify one.)

A notable editorial change is improving the definitions and domintro documentation for the WorkerOptions dictionary members.