Editorial: clarify worker lifetime definitions
The previous definitions relied on long prose sentences with lots of conjunctions and disjunctions. Change them to algorithms that return true or false. These algorithms also operate on the well-defined concept of a WorkerGlobalScope object, instead of the ill-defined concept of a "worker".

Rearrange the order to be actively needed, then protected, then permissible, since these are ever-widening definitions.

Add a clarifying note summarizing how these terms are used, and their relationship to each other, and their relationship to bfcache.

Add a definition, and clarifying note, for the "between-loads shared worker timeout", to make it more prominent.