Editorial: collapse "actively needed" and "protected" for workers
It turns out these concepts are identical, and it was just very hard to tell. Proof:

* "protected" implies "actively needed": immediate from their existing definitions.

* "actively needed" implies "protected": the key is that for an actively needed worker, "the worker's ports" is never empty, so step 3 will always return true. In particular, for dedicated workers, their implicit MessagePort can never be removed from "the worker's ports", since there is no exposed MessagePort object that could be transferred or close()ed or garbage collected. And shared workers are always protected.

Stated another way, despite the complicated definition of "protected", it is not possible to have a worker that is "actively needed" but not "protected". The extra clause about outstanding timers, database transactions, or network connections is redundant.

This is a follow-up to e9ccb97eb12971eed1533f01838370f035ffcf5b. The text prior to that revision also supports this conclusion.