Editorial: further clarify worker lifetime relationships
After e9ccb97eb12971eed1533f01838370f035ffcf5b, the specification contained an incorrect statement in the non-normative note stating that every worker that is actively needed is protected. This is not correct (see also #11598). Expand the non-normative note with more details and examples on the relationships between actively needed, protected, and permissible.