Editorial: relocate "stopped parsing" step for non-DOM documents
As pointed out in https://github.com/whatwg/html/issues/125#issuecomment-3330023993, it was quite hard to discover how the load event fired on iframes containing error documents or other "inline content that doesn't have a DOM". This was because the step that did this was, for some reason, located outside of the main list of algorithm steps. Update to follow the same pattern as used for media documents, of locating the step inline.