Remove incorrect usages of link options destination
The destination field in link options is initialized to the empty string
and can no longer be nullable. Therefore the null check is no longer
valid, since the destination is always set to a non-null value. Secondly,
we already initialize to the empty string and therefore don't need
to reassign the value to the empty string during prefetch loading.