Improve behavior for parsing option end tags
This moves the steps to clone an option element into a
selectedcontent element from the parsing of an option end tag to a
paragraph which specifies that this should happen any time an option is
popped off the stack of open elements, which will make the algorithm run
even if the author didn't explicitly write a closing option tag.

Fixes #11653.