Deferred same-origin cross-document commit
This specifies a way for authors to defer a same-origin cross-document navigation
commit.

It exposes:
- `navigateEvent.deferCommit(promise)`

Allows a listener to the `navigate` event to defer a cross-document commit, without intercepting it.
Returns a promise that can be used to clean up any intermediate state.

- `navigateEvent.destination.ready`

A boolean getter that returns true when the navigation is expected to be instant, i.e.
when the destination already has a document and that document is not render-blocked
(which by and large would imply prerendering).

Closes #11819