Encourage always-escaping ampersand character. (#11988)
In the example highlighting ambiguities from missing semicolons on named
character references, a "correct" encoding is provided, but that example
makes no mention of the fact that the fragment was ambiguous precisely
because the ampersand wasn't escaped.

This patch adds a clarifying note explaining how this situation is
avoided by always escaping the ampersand.

Co-authored-by: Jon Surrell <jon.surrell@automattic.com>
GitHub-PR: 11988
GitHub-PR-URL: https://github.com/whatwg/html/pull/11988