Move "clean up API method tracker" to the right place
This is a step in "commit an intercepted navigate event" that should be done
when the event is *not* intercepted.

This fixes a bug introduced by #11952, where accidentally removing an "Otherwise"
results in calling "clean up" also for same-document navigations, and a subsequent
refactor in #11987 that makes this more nonsensical.